+++ CHANGELOG.md
@@ -1,3 +1,50 @@
+## [1.8.1]
+
+- fix: ensure that Google login on Android always opens in external browser [#455](https://github.com/supabase/supabase-flutter/pull/455)
+
+## [1.8.0]
+
+- feat: allow `signInWithOAuth` with `platformDefault` option to open in app webview for iOS [#431](https://github.com/supabase/supabase-flutter/pull/431)
+
+## [1.7.0]
+
+- feat: Add signInWithApple method [#437](https://github.com/supabase/supabase-flutter/pull/437)
+  ```dart
+  final AuthResponse response = await supabase.auth.signInWithApple();
+  ```
+- fix: upgrade webview_flutter to 4.0.0 [#427](https://github.com/supabase/supabase-flutter/pull/427)
+
+## [1.6.2]
+
+- fix: persist session to local storage on `onAuthStateChanged` event with a session [#422](https://github.com/supabase/supabase-flutter/pull/422)
+
+## [1.6.1]
+
+- fix: log errors from `onAuthStateChange` [#416](https://github.com/supabase/supabase-flutter/pull/416)
+- fix: catch initial session [#418](https://github.com/supabase/supabase-flutter/pull/418)
+
+## [1.6.0]
+
+- feat: Added custom WebView for OAuth (LaunchMode.inAppWebView) [#355](https://github.com/supabase/supabase-flutter/pull/355)
+- fix: update supabase to v1.6.1
+    - add setAuth() function to storage
+    - fix: keep one storage and functions instance to persist auth
+
+## [1.5.0]
+
+- feat: update supabase-dart to 1.6.0 [#381](https://github.com/supabase/supabase-flutter/pull/381)
+  - add support for `signInWithIdToken`
+- fix: declare web support [#392](https://github.com/supabase/supabase-flutter/pull/392)
+
+## [1.4.2]
+
+- fix: show web support on pub.dev [#373](https://github.com/supabase/supabase-flutter/pull/373)
+- refactor: update example [#374](https://github.com/supabase/supabase-flutter/pull/374)
+
+## [1.4.1]
+
+- fix: print stack trace on the console while in debug mode when auth error occurs
+
 ## [1.4.0]
 
 - feat: add `realtimeClientOptions` to `Supabase.initialize()`